Add -Wshorten-64-to-32 to headersclean
Fix existing warnings by casting to the appropriate type.

+++ src/gui/image/qpixmap.h
@@ -163,7 +163,7 @@ inline void QPixmap::scroll(int dx, int dy, int ax, int ay, int awidth, int ahei
 inline bool QPixmap::loadFromData(const QByteArray &buf, const char *format,
                                   Qt::ImageConversionFlags flags)
 {
-    return loadFromData(reinterpret_cast<const uchar *>(buf.constData()), buf.size(), format, flags);
+    return loadFromData(reinterpret_cast<const uchar *>(buf.constData()), uint(buf.size()), format, flags);
 }
